+++ cmd/master/main.go
@@ -1,7 +1,7 @@
 package main
 
 import "fmt"
-import "github.com/aaronang/cong-the-ripper/master"
+import "github.com/aaronang/cong-the-ripper/lib/master"
 
 func main() {
 	fmt.Println("master starting...")
