+++ lib/slave/slave.go
